+++ kernel/src/file/fs/proc/proc_dir/mounts.rs
@@ -57,7 +57,7 @@ impl NodeOps for Mounts {
 		_fs: &dyn Filesystem,
 		off: u64,
 		buf: &mut [u8],
-	) -> EResult<u64> {
+	) -> EResult<usize> {
 		format_content!(off, buf, "{}", self)
 	}
 }
