+++ kernel/src/file/fs/proc/version.rs
@@ -46,7 +46,7 @@ impl NodeOps for Version {
 		_fs: &dyn Filesystem,
 		off: u64,
 		buf: &mut [u8],
-	) -> EResult<u64> {
+	) -> EResult<usize> {
 		format_content!(off, buf, "{} version {}\n", crate::NAME, crate::VERSION)
 	}
 }
